desloppify: split rimap-config/src/validate.rs into focused submodules
Sprint B item 3b. Mechanical decomposition; no behavior change.

validate.rs (1176 LOC) → validate/{mod.rs, identity.rs, limits.rs,
paths.rs, rules.rs}. Production-code portion of mod.rs went from
~439 LOC to ~173 LOC. Tests (737 LOC) stay in mod.rs because they
exercise the full pipeline through the public orchestrator.

  validate/mod.rs (~173 prod + 737 tests)
    - ValidatedAccountConfig + ValidatedMultiConfig public types
    - validate_multi + validate_legacy_as_multi public orchestrators
    - validate_account per-account driver (composes the helpers)
    - All existing unit tests

  validate/identity.rs (44 LOC)
    - validate_imap_username, validate_smtp_username (forward to
      validate_username_field which rejects empty / CR / LF / NUL)
    - parse_fingerprint (TLS pin)

  validate/limits.rs (52 LOC)
    - validate_limits with the ZERO_CHECKS table; new positive-int
      fields get a row instead of another `if` block

  validate/paths.rs (138 LOC)
    - validate_audit_config (retention sanity)
    - validate_paths_multi (writable-dir probes for audit + downloads)
    - default_audit_base / enforce_audit_containment / require_writable_dir

  validate/rules.rs (72 LOC)
    - validate_folder_safety (protected vs expunge conflict check)
    - validate_smtp_required + validate_smtp_encryption
    - resolve_tool_overrides (parse v1 ToolName from string keys)

Helpers are pub(super) so mod.rs can call them; cross-submodule access
goes through mod.rs (the orchestrator), keeping each submodule a leaf.
